dynamically set which "metadata" should be exported by Export course

+++ src/components/web/course-header.tsx
@@ -18,12 +18,13 @@ import { PAGINATION_DEFAULTS } from '#/schemas/search-params'
 import type { TrainerDisplay } from './trainer-manager'
 import { TrainerManager } from './trainer-manager'
 import ExportCourseDialog from '../export-course-dialog'
+import type { ExportMdFileSchema } from '#/schemas/export-file'
 
 interface CourseHeaderProps {
   course: Omit<CourseHeaderData, 'createdAt' | 'updatedAt'>
   variant?: 'default' | 'compact'
   singleCourse?: boolean
-  onExport?: (id: string) => void
+  onExport?: (data: ExportMdFileSchema) => void
   onDelete?: (id: string) => void
   className?: string
   activeTagIds?: string[]
@@ -33,8 +34,8 @@ const CourseHeader = ({
   course,
   variant = 'default',
   singleCourse = true,
-  onExport = async () => {},
-  onDelete = async () => {},
+  onExport,
+  onDelete,
   className,
   activeTagIds = [],
 }: CourseHeaderProps) => {
@@ -73,6 +74,16 @@ const CourseHeader = ({
     }),
   )
   const trainerSize = singleCourse ? 'default' : 'sm'
+
+  const handleExport = (data: ExportMdFileSchema) => {
+    const params = data
+
+    // console.log('CourseHeader, handleExport, data:', params)
+    if (!onExport) return
+    startExportTransition(async () => {
+      await onExport(params)
+    })
+  }
 
   if (variant === 'compact') {
     return (
@@ -157,12 +168,8 @@ const CourseHeader = ({
 
       <CardFooter className="flex flex-row gap-4">
         <ExportCourseDialog
-          type="button"
-          onClick={() => {
-            startExportTransition(async () => {
-              await onExport(course.id)
-            })
-          }}
+          courseId={course.id}
+          onExportSubmit={handleExport}
           disabled={isPending}
           className="hover:cursor-pointer"
         >
@@ -182,11 +189,12 @@ const CourseHeader = ({
         <Button
           type="button"
           variant="destructive"
-          onClick={() =>
+          onClick={() => {
+            if (!onDelete) return
             startDeleteTransition(async () => {
               await onDelete(course.id)
             })
-          }
+          }}
           disabled={isPending}
           className="hover:cursor-pointer"
         >
